requirements.txt
annotated-types==0.7.0
anthropic==0.66.0
anyio==4.10.0
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
coloredlogs==15.0.1
cryptography==45.0.7
defusedxml==0.7.1
distro==1.9.0
fastapi==0.116.1
fastapi_poe==0.0.70
flatbuffers==25.2.10
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
humanfriendly==10.0
idna==3.10
jiter==0.10.0
magika==0.6.2
markdownify==1.2.0
markitdown==0.1.3
mpmath==1.3.0
numpy==2.3.2
onnxruntime==1.22.1
openai==1.106.1
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
protobuf==6.32.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pypdfium2==4.30.0
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sse-starlette==3.0.2
starlette==0.47.3
sympy==1.14.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
